Updated README.md
Added Syntax information for the `generatewallet`, `generateaddress`,
`decryptwallet`, and `decryptwallet filter` sub-commands.

Cleaned up a few things and changed come of the format for better
readability.

The content for the `Command-Line Interface` and `Usage Examples`
sections of `Usage Documentation` are no longer automatically shown
and are now inside of a spoiler.