Security update for requirements.txt
In response to recent security findings, the `requirements.txt` file for the Denaro wallet client has been updated to fortify against vulnerabilities identified in the `pycryptodome` and `Pillow` libraries. 

These vulnerabilities do not directly impact the Denaro wallet client because the vulnerable code segments are not utilized. This update is a preventive measure.

- The `PyCryptodome` library requirement upgraded to version 3.19.1 addresses the vulnerability identified as `CVE-2023-52323`, which pertains to a side-channel leakage during OAEP decryption in both PyCryptodome and pycryptodomex.

- The `Pillow` library requirement upgraded to version 10.2.0 mitigates the risk posed by `CVE-2023-50447`, which allows for Arbitrary Code Execution within Pillow.